Update with PR feedback
